--- src/domain/board/index.ts
@@ -1,2 +1,0 @@
-export { BoardPage } from './pages/BoardPage.tsx';
-export type { BoardPost } from './types.ts';
