+++ src/domain/board/types/index.ts
@@ -1,0 +1,2 @@
+export { BoardPage } from '../pages/BoardPage.tsx';
+export type { BoardPost } from './types.ts';
