+++ src/domain/board/types/types.ts
