+++ src/domain/ruleBook/components/card/styles/card/card.css.ts
