+++ src/domain/ruleBook/components/card/styles/cardContainer/cardContainer.css.ts
