+++ src/domain/ruleBook/styles/ruleBook/guideGlossaryPage.css.ts
