+++ src/domain/ruleBook/styles/ruleBookDetail/guideDetailPage.css.ts
